miration to JUCE 6

+++ JUCE/examples/Utilities/AnalyticsCollectionDemo.h
@@ -2,7 +2,7 @@
   ==============================================================================
 
    This file is part of the JUCE examples.
-   Copyright (c) 2017 - ROLI Ltd.
+   Copyright (c) 2020 - Raw Material Software Limited
 
    The code included in this file is provided under the terms of the ISC license
    http://www.isc.org/downloads/software-support-policy/isc-license. Permission
@@ -31,7 +31,7 @@
 
  dependencies:     juce_analytics, juce_core, juce_data_structures, juce_events,
                    juce_graphics, juce_gui_basics
- exporters:        xcode_mac, vs2017, linux_make, xcode_iphone, androidstudio
+ exporters:        xcode_mac, vs2019, linux_make, xcode_iphone, androidstudio
 
  moduleFlags:      JUCE_STRICT_REFCOUNTEDPOINTER=1
 
@@ -89,7 +89,7 @@ public:
         startAnalyticsThread (initialPeriodMs);
     }
 
-    ~GoogleAnalyticsDestination()
+    ~GoogleAnalyticsDestination() override
     {
         // Here we sleep so that our background thread has a chance to send the
         // last lot of batched events. Be careful - if your app takes too long to
@@ -212,11 +212,10 @@ private:
         // a binary format may be more suitable if it is faster - remember that this
         // method is called on app shutdown so it needs to complete quickly!
 
-        XmlDocument previouslySavedEvents (savedEventsFile);
-        std::unique_ptr<XmlElement> xml (previouslySavedEvents.getDocumentElement());
+        auto xml = parseXMLIfTagMatches (savedEventsFile, "events");
 
-        if (xml.get() == nullptr || xml->getTagName() != "events")
-            xml.reset (new XmlElement ("events"));
+        if (xml == nullptr)
+            xml = std::make_unique<XmlElement> ("events");
 
         for (auto& event : eventsToSave)
         {
@@ -243,50 +242,47 @@ private:
             xml->addChildElement (xmlEvent);
         }
 
-        xml->writeToFile (savedEventsFile, {});
+        xml->writeTo (savedEventsFile, {});
     }
 
     void restoreUnloggedEvents (std::deque<AnalyticsEvent>& restoredEventQueue) override
     {
-        XmlDocument savedEvents (savedEventsFile);
-        std::unique_ptr<XmlElement> xml (savedEvents.getDocumentElement());
-
-        if (xml.get() == nullptr || xml->getTagName() != "events")
-            return;
-
-        auto numEvents = xml->getNumChildElements();
-
-        for (auto iEvent = 0; iEvent < numEvents; ++iEvent)
+        if (auto xml = parseXMLIfTagMatches (savedEventsFile, "events"))
         {
-            auto* xmlEvent = xml->getChildElement (iEvent);
+            auto numEvents = xml->getNumChildElements();
 
-            StringPairArray parameters;
-            auto* xmlParameters = xmlEvent->getChildByName ("parameters");
-            auto numParameters = xmlParameters->getNumAttributes();
+            for (auto iEvent = 0; iEvent < numEvents; ++iEvent)
+            {
+                auto* xmlEvent = xml->getChildElement (iEvent);
 
-            for (auto iParam = 0; iParam < numParameters; ++iParam)
-                parameters.set (xmlParameters->getAttributeName (iParam),
-                                xmlParameters->getAttributeValue (iParam));
+                StringPairArray parameters;
+                auto* xmlParameters = xmlEvent->getChildByName ("parameters");
+                auto numParameters = xmlParameters->getNumAttributes();
 
-            StringPairArray userProperties;
-            auto* xmlUserProperties = xmlEvent->getChildByName ("user_properties");
-            auto numUserProperties = xmlUserProperties->getNumAttributes();
+                for (auto iParam = 0; iParam < numParameters; ++iParam)
+                    parameters.set (xmlParameters->getAttributeName (iParam),
+                                    xmlParameters->getAttributeValue (iParam));
 
-            for (auto iProp = 0; iProp < numUserProperties; ++iProp)
-                userProperties.set (xmlUserProperties->getAttributeName (iProp),
-                                    xmlUserProperties->getAttributeValue (iProp));
+                StringPairArray userProperties;
+                auto* xmlUserProperties = xmlEvent->getChildByName ("user_properties");
+                auto numUserProperties = xmlUserProperties->getNumAttributes();
 
-            restoredEventQueue.push_back ({
-                xmlEvent->getStringAttribute ("name"),
-                xmlEvent->getIntAttribute ("type"),
-                static_cast<uint32> (xmlEvent->getIntAttribute ("timestamp")),
-                parameters,
-                xmlEvent->getStringAttribute ("user_id"),
-                userProperties
-            });
+                for (auto iProp = 0; iProp < numUserProperties; ++iProp)
+                    userProperties.set (xmlUserProperties->getAttributeName (iProp),
+                                        xmlUserProperties->getAttributeValue (iProp));
+
+                restoredEventQueue.push_back ({
+                    xmlEvent->getStringAttribute ("name"),
+                    xmlEvent->getIntAttribute ("type"),
+                    static_cast<uint32> (xmlEvent->getIntAttribute ("timestamp")),
+                    parameters,
+                    xmlEvent->getStringAttribute ("user_id"),
+                    userProperties
+                });
+            }
+
+            savedEventsFile.deleteFile();
         }
-
-        savedEventsFile.deleteFile();
     }
 
     const int initialPeriodMs = 1000;
@@ -336,7 +332,7 @@ public:
         logEventButtonPress.reset (new ButtonTracker (eventButton, "button_press", logButtonPressParameters));
     }
 
-    ~AnalyticsCollectionDemo()
+    ~AnalyticsCollectionDemo() override
     {
         // The event type here should probably be DemoAnalyticsEventTypes::sessionEnd
         // in a more advanced app.
